- fixed: correct Bully priority calculation for all nodes depending on the topology - fixed: use "long" instead of "float" as type for Bully priorities - fixed: for Bully priorities, use additive offset of 100 instead of multiplications by 10 in case a neighbor is found - fixed: Bully priority initialization - added: renamed HRMController::getClusters() to HRMController::getRoutingTargetClusters() - added: HRMController::getRoutingTargets() - fixed: reduced debug output in HRMController - code cleanups - documentation fixes & updates - fixed: simplified HRMConfig, removed "INHERIT_L0_PRIORITY_TO_HIGHER_LEVELS"

+++ fog.routing.hrm/src/de/tuilmenau/ics/fog/packets/election/BullyAnnounce.java
@@ -26,7 +26,7 @@ public class BullyAnnounce extends SignalingMessageBully
 	 * @param pCoordinatorSignature is the signature of the coordinator - can be replaced by cryptographic identity
 	 * @param pToken is the active token that is used for the identification of the domain the coordinator is active in case no Cluster IDs can be provided a priori
 	 */
-	public BullyAnnounce(Name pSenderName, float pSenderPriority, HRMSignature pCoordinatorSignature, int pToken)
+	public BullyAnnounce(Name pSenderName, long pSenderPriority, HRMSignature pCoordinatorSignature, int pToken)
 	{
 		super(pSenderName, pSenderPriority);
 		mCoordSignature = pCoordinatorSignature;
